LayoutTests/legacy-animation-engine/animations/trigger-computed-style-expected.txt

Test the computed style of the -webkit-animation-trigger property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Auto : auto
PASS triggerStyle.length is 1
PASS subRule.cssText is 'auto'

One container-scroll value : container-scroll(10px)
PASS triggerStyle.length is 1
PASS subRule.cssText is 'container-scroll(10px)'

One container-scroll value : container-scroll(2em)
PASS triggerStyle.length is 1
PASS subRule.cssText is 'container-scroll(32px)'

One container-scroll value with end value : container-scroll(10px, 20px)
PASS triggerStyle.length is 1
PASS subRule.cssText is 'container-scroll(10px, 20px)'

Two container-scroll values : container-scroll(10px), container-scroll(20px)
PASS triggerStyle.length is 2
PASS subRule.cssText is 'container-scroll(10px)'
PASS subRule.cssText is 'container-scroll(20px)'

Many container-scroll values : container-scroll(10px), container-scroll(20px), container-scroll(30px), container-scroll(40px), container-scroll(1px)
PASS triggerStyle.length is 5
PASS subRule.cssText is 'container-scroll(10px)'
PASS subRule.cssText is 'container-scroll(20px)'
PASS subRule.cssText is 'container-scroll(30px)'
PASS subRule.cssText is 'container-scroll(40px)'
PASS subRule.cssText is 'container-scroll(1px)'

No value : 
PASS triggerStyle.length is 1
PASS subRule.cssText is 'auto'

Bad value : banana
PASS triggerStyle.length is 1
PASS subRule.cssText is 'auto'

Multiple bad values : banana, eggs, bacon
PASS triggerStyle.length is 1
PASS subRule.cssText is 'auto'

Some good, some bad values : container-scroll(10), eggs, bacon
PASS triggerStyle.length is 1
PASS subRule.cssText is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
